Move save/restore logic to modal-container
