+++ src/ggrc-client/js/templates/issues/modal-content.stache
@@ -3,6 +3,10 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 }}
 
+<modal-container
+  instance:from="instance"
+  model:from="model"
+  isModalSaving:from="isSaving">
 <div class="hideable-holder">
   {{#instance}}
     <issue-main-content-wrapper instance:from="instance" isIssueLinked:from="instance.issueLinked">
@@ -207,3 +211,4 @@
     </issue-main-content-wrapper>
   {{/instance}}
 </div>
+</modal-container>
