+++ src/ggrc-client/js/templates/task_groups/modal-content.stache
@@ -3,6 +3,10 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 }}
 
+<modal-container
+  instance:from="instance"
+  model:from="model"
+  isModalSaving:from="isSaving">
 <form action="javascript://">
   {{> base_objects/form-restore.stache}}
 
@@ -97,3 +101,4 @@
     </div>
   {{/if}}
 </form>
+</modal-container>
